Implement a batch-loading framework for GraphQL queries
We currently perform batch-loading manually via the batch-loader
gem[1], but this is quite cumbersome because it requires a lot of
boilerplate code that is almost always the same for all queries.

The batch-loading framework allows us to batch-load with a single
line of code, and still provides all the flexibility we need.

[1]: https://github.com/exAspArk/batch-loader