Allow the use of `context` in integration tests
For the execution of the query, a shared example is used to which is
add the possibility to pass also the `context` parm (which by default
is nil).
With this change, it's possible to indicate it in the integration
tests which is passed to the shared example which takes care of
triggering the execution of the query with the `context`.