Create an object to create the schema context
This commit moves the responsibility to create the schema context from
the controller to the SolidusGraphqlApi::Context class. It accepts the
request as an argument and implements the `#to_h` method that returns
the context needed by SolidusGraphqlApi::Schema#execute.
For now, includes only the current user but will be integrated with more
context needed stuff.

It also aligns the controller specs with the new implementation.