Add order_token to the schema context
For allowing guests to manage their cart and place their order,
you can use the order's guest token.
This token is contained in the guest_token property of the order,
and it allows you to perform certain checkout-related operations on
the order such as managing line items, completing the checkout flow etc.

The order token must be passed in the X-Spree-Order-Token header in
the following form:

X-Spree-Order-Token: ORDER_TOKEN