Add RSpec aliases, shared_context and metadata
The aliases are 'describe_query', 'connection_field' and 'field'.
They can be used in the specs to explain better the result of a
Graphql query, and they are of type "graphql_query".
This type sets the subject by default through the "graphql query"
shared context.
The subject query can be set using a "let(:query)" or with the "query"
metadata, and also the 'context' and 'variables' can be set in the same
way.
It also adds the "freeze_date" metadata that freeze the date with
Timecop, if set to false the date will be the one set with the
"DEFAULT_FREEZE_DATE" constant, instead if it is a String will be set
by the given date parsed with "DateTime#parse".