Disable codeclimate's rubocop plugin
As stated here[1], Code Climate plugins run in isolated environments and
can't install gems or use gems installed for the repositories they're
analyzing. This is a security feature that does unfortunately limit the
flexibility of plugins to use our own third party plugins.

For this reasons, the RuboCop checks will be moved to the CI.

[1] https://github.com/codeclimate/codeclimate-rubocop/issues/131#issuecomment-416256990