Add RSpec mutation aliases, shared_context and metadata
It is of type "graphql_mutation" and can be used in the specs to explain
better the result of a Graphql mutation.
The "graphql_mutation" sets the subject by default through the
"graphql mutation subject" shared context.
The subject mutation can be set using a "let(:mutation)" or with the
"mutation" metadata, and also the 'context' and 'variables' can be used
in the same way.