Add mutation to remove address from the address book
The mutation accepts an address id as input and removes the address from
the current user address book.

The raises of the CanCan::AccessDenied exception is needed because
Solidus lacks some CanCanCan authorization rules and, in this case,
the rule that is missing is the one needed to authorize the address
deletion.