Add mutation to mark the default address
The mutation accepts an address id as input and sets the address
as the default used in the checkout.

The raises of the CanCan::AccessDenied exceptions are needed because
Solidus lacks some CanCanCan authorization rules and, in this case,
the rule that is missing is the one needed to authorize the address
updates.