Fix `Duplicate type definition ...` error
After a runtime change during the autoloading of the classes
in the development environment (where we usually have the
cache_classes config param set to false), there is a conflict
with the DSL:
At this point, Rails re-loads the files but the code still
retains references to old objects and causes a circular
dependency error in the PaymentSource interface during the
`orphan_type` assignment.

With this change in the `payment_sources` configuration uses
strings to force the correct autoloading of all the classes
which are re-converted into constants during the `orphan_type`
assignment.