Add mutation to add addresses to the current order
The mutation accepts billing and shipping address as input and sets it
for the current order. In case the shipToBillingAddress argument is true,
it ignores the shipping address and instead uses the billing address.

It raises a CanCan::AccessDenied exception in case the current ability
can't update the current order.