Specify store mail_from_address in spec
The store factory was updated in https://github.com/solidusio/solidus/pull/3849

Although I see the "with_args" helper is being used elsewhere to
allow for asserting against dynamic payloads, the only reason
this value is changing on us is because of the unspoken dependency
on the attributes in a solidus factory. If Solidus is going to
embrace this kind of churn then we'd probably be best to set
all relevant attributes explicitly.