Fix duplication of records on has_many_through associations
When a batch included two records, both of them associated to the same
target record through a `has_many_through` association, the target
record was duplicated for both parent records.

We change the name of the models used in the test to reproduce an
example with a clearer business use-case behind.

Fixes #172